requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
scipy>=1.10
tensorflow>=2.12
xgboost>=2.0
pyyaml>=6.0
joblib>=1.3
